Reorganizes Validation Files
