+++ apis/v1alpha2/validation/httproute_validation_test.go
@@ -17,152 +17,12 @@ limitations under the License.
 package validation
 
 import (
-	"reflect"
 	"testing"
 
 	gatewayv1a2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/validation/field"
 	utilpointer "k8s.io/utils/pointer"
 )
-
-func TestValidateGateway(t *testing.T) {
-	listeners := []gatewayv1a2.Listener{
-		{
-			Hostname: nil,
-		},
-	}
-	baseGateway := gatewayv1a2.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "foo",
-			Namespace: metav1.NamespaceDefault,
-		},
-		Spec: gatewayv1a2.GatewaySpec{
-			GatewayClassName: "foo",
-			Listeners:        listeners,
-		},
-	}
-
-	testCases := map[string]struct {
-		mutate             func(gw *gatewayv1a2.Gateway)
-		expectErrsOnFields []string
-	}{
-		"nil hostname": {
-			mutate:             func(gw *gatewayv1a2.Gateway) {},
-			expectErrsOnFields: []string{},
-		},
-		"empty string hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{},
-		},
-		"wildcard hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("*")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{},
-		},
-		"wildcard-prefixed hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("*.example.com")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{},
-		},
-		"valid dns subdomain": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("foo.example.com")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{},
-		},
-		// Invalid use cases
-		"IPv4 address hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("1.2.3.4")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"Invalid IPv4 address hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("1.2.3..4")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"IPv4 address with port hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("1.2.3.4:8080")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"IPv6 address hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("2001:db8::68")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname", "spec.listeners[0].hostname"},
-		},
-		"IPv6 link-local address hostname": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("fe80::/10")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"dns subdomain with port": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("foo.example.com:8080")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"dns subdomain with invalid wildcard label": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("*.*.com")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"dns subdomain with multiple wildcards": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("*.foo.*.com")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-		"dns subdomain with wildcard root label": {
-			mutate: func(gw *gatewayv1a2.Gateway) {
-				hostname := gatewayv1a2.Hostname("*.foo.*.com")
-				gw.Spec.Listeners[0].Hostname = &hostname
-			},
-			expectErrsOnFields: []string{"spec.listeners[0].hostname"},
-		},
-	}
-
-	for name, tc := range testCases {
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			gw := baseGateway.DeepCopy()
-			tc.mutate(gw)
-			errs := ValidateGateway(gw)
-			if len(tc.expectErrsOnFields) != len(errs) {
-				t.Fatalf("Expected %d errors, got %d errors: %v", len(tc.expectErrsOnFields), len(errs), errs)
-			}
-			for i, err := range errs {
-				if err.Field != tc.expectErrsOnFields[i] {
-					t.Errorf("Expected error on field: %s, got: %s", tc.expectErrsOnFields[i], err.Error())
-				}
-			}
-		})
-	}
-}
 
 func TestValidateHTTPRoute(t *testing.T) {
 	testService := "test-service"
@@ -425,77 +285,4 @@ func pathMatchTypePtr(s string) *gatewayv1a2.PathMatchType {
 func portNumberPtr(p int) *gatewayv1a2.PortNumber {
 	result := gatewayv1a2.PortNumber(p)
 	return &result
-}
-
-func TestValidateGatewayClassUpdate(t *testing.T) {
-	type args struct {
-		oldClass *gatewayv1a2.GatewayClass
-		newClass *gatewayv1a2.GatewayClass
-	}
-	tests := []struct {
-		name string
-		args args
-		want field.ErrorList
-	}{
-		{
-			name: "changing parameters reference is allowed",
-			args: args{
-				oldClass: &gatewayv1a2.GatewayClass{
-					Spec: gatewayv1a2.GatewayClassSpec{
-						Controller: "foo",
-					},
-				},
-				newClass: &gatewayv1a2.GatewayClass{
-					Spec: gatewayv1a2.GatewayClassSpec{
-						Controller: "foo",
-						ParametersRef: &gatewayv1a2.ParametersReference{
-							Group: "example.com",
-							Kind:  "GatewayClassConfig",
-							Name:  "foo",
-						},
-					},
-				},
-			},
-			want: nil,
-		},
-		{
-			name: "changing controller field results in an error",
-			args: args{
-				oldClass: &gatewayv1a2.GatewayClass{
-					Spec: gatewayv1a2.GatewayClassSpec{
-						Controller: "foo",
-					},
-				},
-				newClass: &gatewayv1a2.GatewayClass{
-					Spec: gatewayv1a2.GatewayClassSpec{
-						Controller: "bar",
-					},
-				},
-			},
-			want: field.ErrorList{
-				{
-					Type:     field.ErrorTypeInvalid,
-					Field:    "spec.controller",
-					Detail:   "cannot update an immutable field",
-					BadValue: "bar",
-				},
-			},
-		},
-		{
-			name: "nil input result in no errors",
-			args: args{
-				oldClass: nil,
-				newClass: nil,
-			},
-			want: nil,
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			if got := ValidateGatewayClassUpdate(tt.args.oldClass, tt.args.newClass); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("ValidateGatewayClassUpdate() = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
